Replace unicode U+276F with plaintext character
Some users might not have access to unicode characters such as U+276F
which is used as an indicator during fuzzy searching. This could be
replaced with a plain text alternative such as a larger than. For
example this is an option provided by other projects including the
starship prompt.

I've replaced replaced the hardcoded U+276F character with a plain text
larger than.In the future, this could potentially be set by the user in
a config.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>